requirements.txt
Django==4.2.9
djangorestframework==3.14.0
django-environ==0.10.0
mysqlclient==2.2.0
celery==5.3.1
redis==4.6.0
python-dotenv==1.0.1
djangorestframework-simplejwt==2.8.0
drf-spectacular==0.27.1
requests==2.32.3
pandas==2.2.2
loguru==0.7.2
twilio==8.4.0
python-dateutil==2.9.0
